+++ web/src/client/js/reducers/projects.js
@@ -18,17 +18,3 @@ export const projects = (state = initialState, action) => {
       return state
   }
 }
-
-export const project = (state = initialState, action) => {
-  switch (action.type) {
-    // Project
-    case ActionTypes.REQUEST_PROJECT:
-      console.info('Requesting project', action.id)
-      return state
-      break
-    case ActionTypes.RECEIVE_PROJECT:
-      return action.data
-    default:
-      return state
-  }
-}
